services/tts/pyproject.toml
[project]
name = "vtuber-tts"
version = "0.4.0"
description = "Text to speech microservice using Fish Audio"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "httpx>=0.27",
    "pydantic>=2.8",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "ormsgpack>=1.4.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "prometheus-client>=0.19.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
]

[project.optional-dependencies]
fish = [
    "transformers>=4.40",
    "torch",
    "soundfile",
    "numpy",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
